requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
pyperclip>=1.8.2
PyQt6>=6.5.0
webrtcvad>=2.0.10
httpx>=0.25.0
ollama>=0.1.0
google-genai>=1.0.0
assemblyai>=0.20.0
websocket-client>=1.6.0
nvidia-riva-client>=2.25.0
requests>=2.31.0